create gym folder in envpool/mujoco
