+++ envpool/mujoco/gym/inverted_pendulum.h
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-#ifndef ENVPOOL_MUJOCO_INVERTED_PENDULUM_H_
-#define ENVPOOL_MUJOCO_INVERTED_PENDULUM_H_
+#ifndef ENVPOOL_MUJOCO_GYM_INVERTED_PENDULUM_H_
+#define ENVPOOL_MUJOCO_GYM_INVERTED_PENDULUM_H_
 
 #include <algorithm>
 #include <limits>
@@ -24,7 +24,7 @@
 
 #include "envpool/core/async_envpool.h"
 #include "envpool/core/env.h"
-#include "envpool/mujoco/mujoco_env.h"
+#include "envpool/mujoco/gym/mujoco_env.h"
 
 namespace mujoco {
 
@@ -65,10 +65,10 @@ class InvertedPendulumEnv : public Env<InvertedPendulumEnvSpec>,
  public:
   InvertedPendulumEnv(const Spec& spec, int env_id)
       : Env<InvertedPendulumEnvSpec>(spec, env_id),
-        MujocoEnv(
-            spec.config["base_path"_] + "/mujoco/assets/inverted_pendulum.xml",
-            spec.config["frame_skip"_], spec.config["post_constraint"_],
-            spec.config["max_episode_steps"_]),
+        MujocoEnv(spec.config["base_path"_] +
+                      "/mujoco/assets_gym/inverted_pendulum.xml",
+                  spec.config["frame_skip"_], spec.config["post_constraint"_],
+                  spec.config["max_episode_steps"_]),
         healthy_reward_(spec.config["healthy_reward"_]),
         healthy_z_min_(spec.config["healthy_z_min"_]),
         healthy_z_max_(spec.config["healthy_z_max"_]),
@@ -144,4 +144,4 @@ using InvertedPendulumEnvPool = AsyncEnvPool<InvertedPendulumEnv>;
 
 }  // namespace mujoco
 
-#endif  // ENVPOOL_MUJOCO_INVERTED_PENDULUM_H_
+#endif  // ENVPOOL_MUJOCO_GYM_INVERTED_PENDULUM_H_
