+++ envpool/mujoco/gym/pusher.h
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-#ifndef ENVPOOL_MUJOCO_PUSHER_H_
-#define ENVPOOL_MUJOCO_PUSHER_H_
+#ifndef ENVPOOL_MUJOCO_GYM_PUSHER_H_
+#define ENVPOOL_MUJOCO_GYM_PUSHER_H_
 
 #include <algorithm>
 #include <limits>
@@ -24,7 +24,7 @@
 
 #include "envpool/core/async_envpool.h"
 #include "envpool/core/env.h"
-#include "envpool/mujoco/mujoco_env.h"
+#include "envpool/mujoco/gym/mujoco_env.h"
 
 namespace mujoco {
 
@@ -68,7 +68,7 @@ class PusherEnv : public Env<PusherEnvSpec>, public MujocoEnv {
  public:
   PusherEnv(const Spec& spec, int env_id)
       : Env<PusherEnvSpec>(spec, env_id),
-        MujocoEnv(spec.config["base_path"_] + "/mujoco/assets/pusher.xml",
+        MujocoEnv(spec.config["base_path"_] + "/mujoco/assets_gym/pusher.xml",
                   spec.config["frame_skip"_], spec.config["post_constraint"_],
                   spec.config["max_episode_steps"_]),
         ctrl_cost_weight_(spec.config["ctrl_cost_weight"_]),
@@ -169,4 +169,4 @@ using PusherEnvPool = AsyncEnvPool<PusherEnv>;
 
 }  // namespace mujoco
 
-#endif  // ENVPOOL_MUJOCO_PUSHER_H_
+#endif  // ENVPOOL_MUJOCO_GYM_PUSHER_H_
