+++ setup.cfg
@@ -41,7 +41,7 @@ envpool = atari/*.so
     vizdoom/*.so
     mujoco/*.so
     mujoco/*.so.*
-    mujoco/assets/*
+    mujoco/assets*/*
 
 [yapf]
 based_on_style = yapf
